refactor: rename to TextEditor and move up one level
